requirements/base.txt
pytz==2021.1  # https://github.com/stub42/pytz
python-slugify==4.0.1  # https://github.com/un33k/python-slugify
Pillow==8.1.0  # https://github.com/python-pillow/Pillow
argon2-cffi==20.1.0  # https://github.com/hynek/argon2_cffi
whitenoise==5.2.0  # https://github.com/evansd/whitenoise
redis==3.5.3  # https://github.com/andymccurdy/redis-py
hiredis==1.1.0  # https://github.com/redis/hiredis-py
celery==4.4.6  # pyup: < 5.0,!=4.4.7  # https://github.com/celery/celery
django-celery-beat==2.2.0  # https://github.com/celery/django-celery-beat
flower==0.9.7  # https://github.com/mher/flower

# Django
# ------------------------------------------------------------------------------
django==3.0.12  # pyup: < 3.1  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==4.1.1  # https://github.com/jazzband/django-model-utils
django-allauth==0.44.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.11.0  # https://github.com/django-crispy-forms/django-crispy-forms
django-redis==4.12.1  # https://github.com/jazzband/django-redis
django-measurement==3.2.3
django-bootstrap-modal-forms==2.0.1
fontawesome-free==5.15.2
django-widget-tweaks==1.4.8
django-filter==2.4.0
django-dynamic-formsets==0.0.8
attrdict==2.0.1
django-weasyprint==1.1.0
python-slugify==4.0.1